Make bitrev tables optional
This introduces a feature "bitrev-tables" that can be used to opt into
the use of bitrev tables for the bit-reverse reordering at the beginning
of an FFT.

Bitrev tables are not disabled by default, because they consume a
signficant amount of memory and actually provide worse performance on
chips that provide a dedicated bit reversal instruction (like `RBIT` on
ARMv7). They are still beneficial on systems that don't provide such an
instruction though.